Capability for Cocoa beans
Added capability of adding Cocoa beans if the blocks surrounding the target block is a log (with id 17) and a jungle wood (with BlockData/meta 3)
added the itemId (127) for cocoa sapling, do not confuse with the cocoa beans item which is 351:3 which is BROWN DYE.
I didnt know if the itemId was the item in the inventory ID or the block ID when the sapling is placed down. 

somebody will have to check that.